requirements.txt
imageio
matplotlib
scikit-image

easydict
glob2
path.py

opencv-python>=3.0,<4.0
tensorboardX
torch @ https://download.pytorch.org/whl/cu100/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu100/torchvision-0.3.0-cp36-cp36m-linux_x86_64.whl
